etf_bot/requirements.txt
python-telegram-bot==21.8
google-genai
requests
python-dotenv